requirements.txt
Django==4.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
redis==4.4.4
walrus==0.9.2
celery==5.2.7
requests==2.31.0
python-dotenv==0.19.2
pyproj==3.3.1
pandas==1.4.3
Shapely==2.0.1
python-jose==3.3.0
cryptography==39.0.1
pyjwt==2.4.0
jwcrypto==1.4.2
geojson==2.5.0
arrow==1.2.2
tldextract==3.2.1
rtree==1.0.1
psycopg2-binary==2.9.3
marshmallow==3.17.0
implicitdict==1.1.1
pika==1.3.1
http-message-signatures==0.4.4
dj-database-url==1.0.0
django-celery-beat==2.5.0
dacite==1.8.1
